@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # of the product and file versions while build-number is used as the build suffix.
 version: 1.1.0+10
 environment:
-  sdk: '>=2.17.1 <3.0.0'
+  sdk: ">=3.6.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,11 +30,12 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  csc_picker: ^0.2.7
-  intl: ^0.17.0
+#  csc_picker: ^0.2.7
+  intl: ^0.20.2
   image_picker: ^1.0.1
   auto_size_text: ^3.0.0
   animated_bottom_navigation_bar: ^1.2.0
@@ -42,9 +43,9 @@ dependencies:
   dio: ^5.3.0
   cached_network_image: ^3.2.3
   fluttertoast: ^8.1.2
-  flutter_bloc: ^8.1.1
+  flutter_bloc: ^9.1.0
   gradient_borders: ^1.0.0
-  carousel_slider: ^4.2.1
+  carousel_slider: ^5.0.0
   shared_preferences: ^2.0.15
   flutter_spinkit: ^5.1.0
   flutter_dialogs: ^3.0.0
@@ -55,7 +56,7 @@ dependencies:
   metadata_extract: ^1.0.0
   any_link_preview: ^3.0.0
   path_provider: ^2.0.11
-  share_plus: ^7.0.2
+  share_plus: ^10.1.4
   flutter_gradient_colors: ^2.1.1
   table_calendar:
   flutter_holo_date_picker: ^1.1.0
@@ -65,21 +66,21 @@ dependencies:
   flutter_contacts:
   twilio_flutter: ^0.0.9
   receive_sharing_intent: ^1.4.5
-  webview_flutter: ^2.0.12
+  webview_flutter: ^4.10.0
   animations: ^2.0.7
   #  webview_flutter_android: ^3.9.2
   #  webview_flutter_wkwebview: ^3.7.2
   socket_io_client:
-  firebase_core: ^2.13.0
-  firebase_auth: ^4.2.9
-  googleapis: ^11.3.0
-  firebase_dynamic_links: ^5.0.15
+  firebase_core: ^3.12.1
+  firebase_auth: ^5.5.1
+  googleapis: ^13.1.0
+  firebase_dynamic_links: ^6.1.4
   mailer: ^6.0.0
   # flutter_html: ^3.0.0-alpha.6
   flutter_html: ^3.0.0-beta.2
   flutter_sms: ^2.3.3
   timezone: ^0.9.1
-  cloud_firestore_web: ^3.3.5
+  cloud_firestore_web: ^4.4.5
   google_sign_in: ^5.0.2
   badges: any
   moment_dart: ^1.0.1
